requirements.txt
numpy
matplotlib
tensorflow
tensorboard
terminaltables
pillow
tqdm